+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 1.1.3
 * Addon now hooks directly to mail functions for improved compability with other addons.
+* Added option to reset history data from today and yesterday.
+* Added tooltips to clarify meaning of some settings.
 
 ## 1.1.2
 * Attempted fix to tooltip nil error in OnClick callback.